Created mesh class
Created a mesh class to handle all of the buffers and OpenGL code completely abstracted from the main file. Also removed the VertexBufferLayout and changed to use a Vertex struct instead. Also added support for automating the creation of uniforms depending on how many textures that is in the Textures vector when creating a new mesh class!